requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas==2.2.0
boto3==1.34.34
requests==2.31.0
pydantic==2.5.3
python-multipart==0.0.6
sqlalchemy==2.0.25
pymysql==1.1.0
psycopg2-binary==2.9.9
simple-salesforce==1.12.6
openpyxl==3.1.2
duckdb==0.10.0
python-dotenv==1.0.0
